refactor: add alias path
