+++ craco.config.js
@@ -1,0 +1,14 @@
+const CracoAlias = require('craco-alias');
+module.exports = {
+  plugins: [
+    {
+      plugin: CracoAlias,
+      options: {
+        source: 'tsconfig',
+        baseUrl: '.',
+        tsConfigPath: 'tsconfig.paths.json',
+        debug: false,
+      },
+    },
+  ],
+};
