+++ package.json
@@ -5,6 +5,8 @@
   "proxy": "https://api.sgsg.space/v1",
   "dependencies": {
     "@chakra-ui/react": "^1.3.4",
+    "@craco/craco": "^6.2.0",
+    "@emotion/babel-plugin-jsx-pragmatic": "^0.1.5",
     "@emotion/react": "^11.1.5",
     "@emotion/styled": "^11.1.5",
     "@testing-library/jest-dom": "^5.11.9",
@@ -19,7 +21,6 @@
     "@types/react-router-dom": "^5.1.7",
     "framer-motion": "^3.10.3",
     "howler": "^2.2.1",
-    "moment": "^2.29.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-moment": "^1.1.1",
@@ -31,9 +32,9 @@
     "web-vitals": "^1.1.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "start": "craco start",
+    "build": "craco build",
+    "test": "craco test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -55,10 +56,14 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-transform-react-jsx": "^7.14.5",
+    "@emotion/babel-plugin": "^11.3.0",
     "@svgr/cli": "^5.5.0",
     "@typescript-eslint/eslint-plugin": "^4.18.0",
     "@typescript-eslint/parser": "^4.18.0",
     "axios": "^0.21.1",
+    "babel-plugin-module-resolver": "^4.1.0",
+    "craco-alias": "^3.0.1",
     "eslint": "^7.22.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-airbnb-base": "^14.2.1",
