+++ src/router/index.tsx
@@ -2,7 +2,7 @@
 // @ts-nocheck
 import React from 'react';
 import { createBrowserHistory } from 'history';
-import { BrowserRouter, Redirect, Route, Switch } from 'react-router-dom';
+import { Router, Redirect, Route, Switch } from 'react-router-dom';
 import Login from '../pages/Login';
 import AfterLogin from '../pages/AfterLogin';
 import PrivateRoute from './PrivateRoute';
@@ -12,9 +12,9 @@ import NotFound from '../pages/NotFound';
 
 const history = createBrowserHistory();
 
-const Router = () => {
+const AppRouter = () => {
   return (
-    <BrowserRouter history={history}>
+    <Router history={history}>
       <Switch>
         <Route exact path="/login" component={Login} />
         <Route exact path="/after-login" component={AfterLogin} />
@@ -23,8 +23,8 @@ const Router = () => {
         <Route path="/error" component={NotFound} />
         <Redirect path="*" to="/" /> *
       </Switch>
-    </BrowserRouter>
+    </Router>
   );
 };
 
-export default Router;
+export default AppRouter;
