+++ tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "extends": "./tsconfig.paths.json",
   "compilerOptions": {
     "target": "es5",
     "lib": [
@@ -24,6 +25,7 @@
     ]
   },
   "include": [
-    "src"
+    "src",
+    "craco.config.js"
   ]
 }
