+++ tsconfig.paths.json
@@ -1,0 +1,7 @@
+{ "compilerOptions":
+  { "baseUrl": ".",
+      "paths": {
+        "@/*": [ "src/*" ]
+      }
+  }
+}
